refactor: changed routes test to a separate file
